NEW: tests.helpers: higher-order functions for standard tests
- as prototyped for the connectivity tests, make these helper routines
available for the tests module
- selection generatorion via itertools
- higher order runners with frontend derived callables as argument

On branch preproc-tests
Changes to be committed:
	new file:   syncopy/tests/helpers.py
	modified:   syncopy/tests/test_connectivity.py
